requirements.txt

opik
fastapi
ipdb
pydantic
uvicorn[standard]
langgraph
langchain-openai
langchain
pydantic_settings
sentence_transformers
faiss-cpu
sentry-sdk[fastapi]